backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
alembic==1.13.1
psycopg2-binary>=2.9.0
redis==5.0.3
celery==5.3.6
influxdb-client==1.42.0
elasticsearch==8.13.0
boto3==1.34.84
PyJWT==2.8.0
cryptography==42.0.5
structlog==24.1.0
Flasgger==0.9.7.1
pydantic==2.12.5
Flask-SocketIO==5.3.6
eventlet==0.36.1
twilio==9.0.4
sendgrid==6.11.0
# AI / Agents
langchain==1.2.13
langchain-core==1.2.23
langchain-google-genai>=2.1.0
langchain-community==0.0.38
langgraph==1.1.3
langgraph-prebuilt==1.0.8
langsmith==0.7.23
pinecone-client==3.2.2
google-generativeai>=0.8.0
google-genai>=1.0.0
groq>=0.11.0
elevenlabs>=1.5.0
google-cloud-texttospeech>=2.17.0
google-auth>=2.29.0
# Auth / Utils
python-dotenv
flask-migrate
flask-jwt-extended==4.6.0
bcrypt==4.1.2
flask-limiter
boto3==1.34.84
PyMuPDF==1.27.2.2
